refactor(kaizen): extract Control Protocol helpers to ControlProtocolMixin
The first extraction put base_agent.py at 1004 against a 1015 limit. That is
11 lines of headroom, and the file had just overshot by 53 through ordinary
feature work — so the guard would have gone red again within a change or two
and the next session would have faced the same raise-the-limit temptation
this issue exists to refuse. Restoring the invariant means restoring its
margin, not clearing it by one line.

ask_user_question / request_approval / report_progress are the natural second
unit: they were already fenced off under their own section banner, and they
share a precondition (a configured control_protocol), a request type
(ControlRequest) and a failure mode (error response -> RuntimeError naming
the interaction). They depend on exactly one piece of host state, which the
mixin documents as its duck-typing contract; __init__ still assigns it.

ControlRequest stays a lazy in-method import rather than moving to module
scope, so an agent that never speaks the Control Protocol still does not pay
to import kaizen.core.autonomy.

Public API unchanged — all three resolve through the MRO. 1004 -> 922, which
puts the file back near the 882 the 1015 limit was derived from.

Fixes #2119
